+++ client/src/redux/actions/GET/a_getBooks.js
@@ -2,30 +2,29 @@
 import axios from 'axios'
 
 // URLS
-import { LOCAL_BE_base_URL } from '../../utils'
-import { LIVE_BE_base_URL } from '../../utils'
-import { GET_BOOKLISTS_START, GET_BOOKLISTS_FAILURE } from './a_lists'
+import { LOCAL_BE_base_URL } from '../../../utils'
+import { LIVE_BE_base_URL } from '../../../utils'
 
 // __MAIN__
     // Action Types
-    export const GET_REVIEWS_START = "GET_REVIEWS_START"
-    export const GET_REVIEWS_SUCCESS = "GET_REVIEWS_SUCCESS"
-    export const GET_REVIEWS_FAILURE = "GET_REVIEWS_FAILURE"
+    export const GET_BOOKS_START = "GET_BOOKS_START"
+    export const GET_BOOKS_SUCCESS = "GET_BOOKS_SUCCESS"
+    export const GET_BOOKS_FAILURE= "GET_BOOKS_FAILURE"
 
-    // Action Creator
-    export const a_getReviews = () => {
-    // console.log('INSIDE: a_getReviews action creator')
-    // -- // 
+    // Action Creator:
+    export const a_getBooks = () => {
+    // console.log('INSIDE: a_GetBooks action creator')
+    // -- //
         // Send First Action
         return dispatch => {
-            dispatch({ type: GET_REVIEWS_START})
+            dispatch({ type: GET_BOOKS_START })
             let used_URL = ''
 
             // What environment are we in?
             if (process.env.NODE_ENV === 'development') {
-                used_URL = `${LOCAL_BE_base_URL}reviews/all`
+                used_URL = `${LOCAL_BE_base_URL}books/all`
             } else if (process.env.NODE_ENV === 'production') {
-                used_URL = `${LIVE_BE_base_URL}reviews/all`
+                used_URL = `${LIVE_BE_base_URL}books/all`
             }
             // console.log('URL USED')
             // console.log(used_URL)
@@ -35,19 +34,19 @@ import { GET_BOOKLISTS_START, GET_BOOKLISTS_FAILURE } from './a_lists'
                 .get(
                     used_URL
                 )
-                .then( res => {
+                .then(res => {
                 // console.log(res)
                 // -- //
                     dispatch({
-                        type: GET_REVIEWS_SUCCESS,
+                        type: GET_BOOKS_SUCCESS,
                         payload: res.data,
                     })
                 })
                 .catch( err => {
                 // console.log(err)
-                // -- // 
+                // -- //
                     dispatch({
-                        type: GET_BOOKLISTS_FAILURE,
+                        type: GET_BOOKS_FAILURE,
                         payload: err
                     })
                 })
