+++ client/src/redux/actions/GET/a_getReviews.js
@@ -2,29 +2,29 @@
 import axios from 'axios'
 
 // URLS
-import { LOCAL_BE_base_URL } from '../../utils'
-import { LIVE_BE_base_URL } from '../../utils'
+import { LOCAL_BE_base_URL } from '../../../utils'
+import { LIVE_BE_base_URL } from '../../../utils'
 
 // __MAIN__
     // Action Types
-    export const GET_BOOKS_START = "GET_BOOKS_START"
-    export const GET_BOOKS_SUCCESS = "GET_BOOKS_SUCCESS"
-    export const GET_BOOKS_FAILURE= "GET_BOOKS_FAILURE"
+    export const GET_REVIEWS_START = "GET_REVIEWS_START"
+    export const GET_REVIEWS_SUCCESS = "GET_REVIEWS_SUCCESS"
+    export const GET_REVIEWS_FAILURE = "GET_REVIEWS_FAILURE"
 
-    // Action Creator:
-    export const a_getBooks = () => {
-    // console.log('INSIDE: a_GetBooks action creator')
-    // -- //
+    // Action Creator
+    export const a_getReviews = () => {
+    // console.log('INSIDE: a_getReviews action creator')
+    // -- // 
         // Send First Action
         return dispatch => {
-            dispatch({ type: GET_BOOKS_START })
+            dispatch({ type: GET_REVIEWS_START})
             let used_URL = ''
 
             // What environment are we in?
             if (process.env.NODE_ENV === 'development') {
-                used_URL = `${LOCAL_BE_base_URL}books/all`
+                used_URL = `${LOCAL_BE_base_URL}reviews/all`
             } else if (process.env.NODE_ENV === 'production') {
-                used_URL = `${LIVE_BE_base_URL}books/all`
+                used_URL = `${LIVE_BE_base_URL}reviews/all`
             }
             // console.log('URL USED')
             // console.log(used_URL)
@@ -34,19 +34,19 @@ import { LIVE_BE_base_URL } from '../../utils'
                 .get(
                     used_URL
                 )
-                .then(res => {
+                .then( res => {
                 // console.log(res)
                 // -- //
                     dispatch({
-                        type: GET_BOOKS_SUCCESS,
+                        type: GET_REVIEWS_SUCCESS,
                         payload: res.data,
                     })
                 })
                 .catch( err => {
                 // console.log(err)
-                // -- //
+                // -- // 
                     dispatch({
-                        type: GET_BOOKS_FAILURE,
+                        type: GET_REVIEWS_FAILURE,
                         payload: err
                     })
                 })
