+++ client/src/redux/actions/POST/a_addAndLogBook.js
@@ -2,24 +2,24 @@
 import axios from 'axios'
 
 // URLS
-import { LOCAL_BE_base_URL } from '../../utils'
-import { LIVE_BE_base_URL } from '../../utils'
+import { LOCAL_BE_base_URL } from '../../../utils'
+import { LIVE_BE_base_URL } from '../../../utils'
 
 // IMPORT Action Types
 // - 1 - //
     import {
         ADD_BOOK_START, ADD_BOOK_SUCCESS, ADD_BOOK_FAILURE
-    } from '../actions/a_addBook.js'
+    } from './a_addBook.js'
 // - 2 - //
     import {
         LOG_COMPLETEDBOOK_START, LOG_COMPLETEDBOOK_SUCCESS, LOG_COMPLETEDBOOK_FAILURE
-    } from '../actions/a_logCompletedBook.js'
+    } from './a_addCompletedBook.js'
 
 // IMPORT Action Creators
 // - 1 - //
-    import { a_addBook } from '../actions/a_addBook.js'
+    import { a_addBook } from './a_addBook.js'
 // - 2 - //
-    import { a_logCompletedBook } from '../actions/a_logCompletedBook.js'
+    import { a_logCompletedBook } from './a_addCompletedBook.js'
 
 // __MAIN__
     // Action Creator
